backend/requirements.txt
# Python version
python-3.13.0

# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.9

# Database & ORM
SQLAlchemy>=2.0.0
psycopg2-binary  # For PostgreSQL on Render

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# AI & LLM
openai>=1.0.0

# File Processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Audio/Video Processing
librosa
opencv-python-headless  # Headless version for servers
numpy

# Payment (Optional - can remove if not using)
python-alipay-sdk
wechatpayv3

# OAuth (Optional)
authlib
httpx

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
redis>=5.0.0
